added read & write runtime permission in splash
added read, write runtime permission and also created directory 'Music_Streaming' in internal storage.
Also enable to access the username and password if the file is available in the internal storage, this will help is the application is deleted and user don't remember his username and password using the file it can make it login again.